properly rename and bucket ages
